# travel-partner-preferences-api/travel-partner-preferences-api-restful/src/main/resources/openapi/preferences-restful-api.yml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Travel Companion Domain
  license:
    name: MIT


tags:
  - name: preferences
    description: Everything about preferences
    externalDocs:
      description: Find out more
      url: http://swagger.io

paths:
  /preferences:
    post:
      tags:
        - preferences
      summary: Add a new prefernces
      description: Add a new preferences
      operationId: addPreferences
      requestBody:
        description: Create a new preferences in store
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PreferencesDto'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PreferencesDto'          
        '405':
          description: Invalid input
    
    put:
      tags:
        - preferences
      summary: Update an existing preferences
      description: Update an existing references by user id
      operationId: updatePreferences
      requestBody:
        description: Update an existent preferences in store
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PreferencesDto'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PreferencesDto'          
        '400':
          description: Invalid ID supplied
        '404':
          description: User id not found
        '405':
          description: Validation exception
          
  /preferences/{userId}:
    delete:
      tags:
        - preferences
      summary: Deletes a preferences by suer id
      description: delete a preferences of the store by user id
      operationId: deletePreferences
      parameters:
        - name: userId
          in: path
          description: user id to delete preferences
          required: true
          schema:
            type: string
      responses:
        '400':
          description: Invalid pet value
    

components:
  schemas:
    PreferencesDto:
      type: object
      additionalProperties: false
      description: >-
        Preference by user.
      required:
        - userId
      properties:
        userId:
          type: string
        description:
          type: string
          
    ErrorDto:
      type: object
      additionalProperties: false
      description: >-
        Common error.
      properties:
        code:
          type: string
        description:
          type: string